Refactoring the repository
We're not splitting sans and sans-client. Why? Because one deals
more with hardware abstractions over the libreflip hardware. This
library can be invoked from multiple binary end-points and as
such, I think it's a good pattern to break common functionality
out into a library crate.
We can even publish it on crates.io if we want to ✨