This includes even more refactoring.
I promise I'll add documentation to the wiki soon (looks at self critically).
Realistically, this is changing the structure of the project because of what
I've learned in the last 4 months about building Rust tools.

There are now three libary crates that are implemented by three tools (two that
are actually MVP and currently "present").

- sans-core
  - sans-types
  - sans-processing

- sans-server
  - sans-types
  - sans-core
- sans-worker
- sans-ctl